requirements.txt
colorama==0.4.3
google_api_python_client==2.133.0
httplib2==0.22.0
protobuf==5.27.1
pyTelegramBotAPI==4.19.1
Telethon==1.36.0
PySocks==1.7.1
